refactor: MiniCalendar 컴포넌트 - 컴포넌트 및 스타일 분리 (#30)
