+++ front/src/components/SideBar/MiniCalendar/index.jsx
@@ -1,9 +1,11 @@
 import React, { useEffect } from 'react';
 import { useState } from 'react';
 import styles from './style.module.css';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'; //폰트어썸
-import { faAngleLeft, faAngleRight } from '@fortawesome/free-solid-svg-icons'; //폰트어썸
-import { getMonth, WEEK_DAYS } from '../../../utils/moment';
+
+import CalendarHeader from './CalendarHeader';
+import WeekDaysHeader from './WeekDaysHeader';
+import Week from './Week';
+import { getMonth } from '../../../utils/moment';
 
 const Index = () => {
   const [selectedDate, setSelectedDate] = useState(new Date());
@@ -13,69 +15,25 @@ const Index = () => {
     setMonth(getMonth(selectedDate));
   }, [selectedDate]);
 
-  function moveMonth(number) {
-    setSelectedDate(preDate => {
-      const date = new Date(preDate.getTime());
-      date.setMonth(date.getMonth() + number);
-      return date;
-    });
-  }
-
-  function initDateClassName(date) {
-    let className = '';
-    if (date.isToday()) className += styles.today;
-    else if (date.month !== selectedDate.getMonth() + 1)
-      className += styles.blur;
-    else if (date.time === selectedDate.getTime())
-      className += styles.button_select;
-
-    return className;
-  }
-
   return (
     <div className={styles.calendar_wrap}>
-      <div className={styles.year}>
-        <div className={styles.text}>
-          <em>{selectedDate.getFullYear()}년</em>
-          <em>{selectedDate.getMonth() + 1}월</em>
-        </div>
-        <div className={styles.year_btt}>
-          <button onClick={() => moveMonth(-1)}>
-            <FontAwesomeIcon icon={faAngleLeft} />
-          </button>
-          <button onClick={() => moveMonth(1)}>
-            <FontAwesomeIcon icon={faAngleRight} />
-          </button>
-        </div>
-      </div>
+      <CalendarHeader
+        selectedDate={selectedDate}
+        setSelectedDate={setSelectedDate}
+      ></CalendarHeader>
       <table className={styles.calendar_talbe}>
         <thead>
-          <tr>
-            {WEEK_DAYS.map(item => {
-              return <th key={item}>{item}</th>;
-            })}
-          </tr>
+          <WeekDaysHeader />
         </thead>
         <tbody>
-          {month.map((week, index) => {
-            return (
-              <tr key={index}>
-                {week.map((date, index) => {
-                  return (
-                    <td
-                      key={index}
-                      className={`${initDateClassName(date)}`}
-                      onClick={() => {
-                        setSelectedDate(new Date(date.time));
-                      }}
-                    >
-                      <em>{date.date}</em>
-                    </td>
-                  );
-                })}
-              </tr>
-            );
-          })}
+          {month.map((week, index) => (
+            <Week
+              key={index}
+              week={week}
+              selectedDate={selectedDate}
+              setSelectedDate={setSelectedDate}
+            />
+          ))}
         </tbody>
       </table>
     </div>
